Attempt to fix it

+++ src/Functions/LeftRight.h
@@ -99,7 +99,7 @@ public:
                 // According to the docs, if length_value < 0, we need to take a suffix of the string starting from the position abs(length_value)
                 else
                 {
-                    sliceFromLeftConstantOffsetUnbounded(source, StringSink(*col_res, input_rows_count), -length_value);
+                    sliceFromLeftConstantOffsetUnbounded(source, StringSink(*col_res, input_rows_count), -static_cast<UInt64>(length_value));
                 }
             }
             else
